% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{chains.sigma}
\alias{chains.sigma}
\title{Create a new sigma based on parameter chain}
\usage{
chains.sigma(chains)
}
\arguments{
\item{chains}{An mcmc.chain or mcmc.chains}
}
\value{
A suggested sigma function for multivariate normal proposal function
}
\description{
Update uses cov(chain) * 2.38^2 / # params
}
