% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{chains_param}
\alias{chains_param}
\title{Get all samples for a given parameter across chains}
\usage{
chains_param(chains, param)
}
\arguments{
\item{chains}{mcmc.chains object}

\item{param}{name of parameter to get samples for}
}
\description{
Get all samples for a given parameter across chains
}
