% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{combine.chains}
\alias{combine.chains}
\title{Get all chain samples from the chains in an mcmc.chains objects as a matrix}
\usage{
combine.chains(chains)
}
\arguments{
\item{chains}{mcmc.chains}
}
\value{
a matrix with each column a parameter
}
\description{
Get all chain samples from the chains in an mcmc.chains objects as a matrix
}
