% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{init.params}
\alias{init.params}
\title{Get initial parameter values that give a finite likelihood}
\usage{
init.params(param.names, ll, chains = 1, prior)
}
\arguments{
\item{param.names}{names of parameters (also indicates the number of parameters)}

\item{ll}{log-likelihood (or psuedo log-likelihood) function}

\item{chains}{if > 1, a list of initial parameter vectors will be returned}
}
\value{
If a non-0 likelihood is found within 1000 uniform random samples, the corresponding parameter vector is returned
}
\description{
Get initial parameter values that give a finite likelihood
}
