% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{init.sig}
\alias{init.sig}
\title{Initialize the std dev of proposal function}
\usage{
init.sig(p.var)
}
\arguments{
\item{p.var}{vector of variances of parameter priors}
}
\value{
a nxn diagonal matrix, with entries sqrt(p.var)/10, or 1/10 the sd of priors
}
\description{
Initialize the std dev of proposal function
}
