% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{lp.jeff}
\alias{lp.jeff}
\title{Log-probability density of logistic transform of jeffery prior}
\usage{
lp.jeff(x)
}
\arguments{
\item{x}{parameter, number or numeric vector}
}
\value{
f(x) for logit(Beta(1/2,1/2)) distribution
}
\description{
Note that for f(x) = logit(Beta(1/2,1/2)), we have E(f(x)) = 0, Var(f(x)) = pi^2
}
