% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{lp.logistic.jeff}
\alias{lp.logistic.jeff}
\title{This is a non-independent Jeffery prior for logistic regression parameters.
See Chen et al, 2008: Properties and Implementation of Jeffreys's Prior in Binomial Regression Models}
\usage{
lp.logistic.jeff(X)
}
\arguments{
\item{X}{model matrix for logistic regression}
}
\value{
a function taking beta, logistic regression parameters, and returning log-prob of jeffery prior (up to additive constant)
}
\description{
This is a non-independent Jeffery prior for logistic regression parameters.
See Chen et al, 2008: Properties and Implementation of Jeffreys's Prior in Binomial Regression Models
}
