% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{lprob_obs_pri}
\alias{lprob_obs_pri}
\title{Probability of observed household outcome given primary cases}
\usage{
lprob_obs_pri(obs, pri, q, Sn = 1, Sp = 1, it = FALSE)
}
\arguments{
\item{obs}{observed household cases (vector) NA=missed}

\item{pri}{primary household infections (vector)}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}

\item{it}{whether to use iteration for secondary cases}
}
\value{
probability of observation
}
\description{
Probability of observed household outcome given primary cases
}
