% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{lprob_sec_obs_pri}
\alias{lprob_sec_obs_pri}
\title{Probability of observed cases and missed observations among potential secondary infections given the number of primary infections.}
\usage{
lprob_sec_obs_pri(obs, mis, pri, m, q, Sn = 1, Sp = 1)
}
\arguments{
\item{obs}{number of observed positive infections among potential secondary infections.}

\item{mis}{number of missed observation among potential secondary infections.}

\item{pri}{number of primary infections, which are potential sources of secondary infection.}

\item{m}{number of potential secondary household infections.}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}
}
\value{
probability of observation
}
\description{
Secondary infections are computed using a generating functions and depend only on the number of secondary infections, independant of individuals infected.
Secondary infection probabilities are computed by iterating over possible numbers of true positives, and counting corresponding potential secondary infections
}
