% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{lprob_sec_obs_pri_it}
\alias{lprob_sec_obs_pri_it}
\title{Probability of observed cases and missed observations among potential secondary infections given the number of primary infections, using iteration for secondary cases.}
\usage{
lprob_sec_obs_pri_it(obs, pri, m, q, Sn = 1, Sp = 1)
}
\arguments{
\item{obs}{vector of observations, -1=missed, 0=negative, 1=positive}

\item{pri}{number of primary infections, which are potential sources of secondary infection.}

\item{m}{number of potential secondary household infections.}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}

\item{mis}{number of missed observation among potential secondary infections.}
}
\value{
probability of observation
}
\description{
Using an iterative method for secondary infections, for performance comparison only
}
\details{
Secondary infections are computed using a generating functions and depend only on the number of secondary infections, independant of individuals infected.
Secondary infection probabilities are computed by iterating over possible numbers of true positives, and counting corresponding potential secondary infections
}
