% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.chain}
\alias{mcmc.chain}
\title{MCMC Function method suitable for psuedo-likelihood}
\usage{
mcmc.chain(n, ll, p0, sigma, lp, psuedo.ll = TRUE)
}
\arguments{
\item{n}{length of chain}

\item{ll}{log likelihood function, taking as arguments parameters p0}

\item{p0}{initial value of parameters}

\item{sigma}{covariance matrix for multivariate gaussian update kernel}

\item{lp}{function of parameters p0 giving the prior log-probability for each parameter}

\item{psuedo.ll}{indicates if ll is psuedo-likelihood, in which case it must be recalculated each time}
}
\value{
mcmc.chain object, containing chain samples, acceptance variables and computed log-likelihoods
}
\description{
MCMC Function method suitable for psuedo-likelihood
}
