% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.chains}
\alias{mcmc.chains}
\title{Build a set of MCMC chains}
\usage{
mcmc.chains(num.chains, n, ll, p0.list, sigma, lp, psuedo.ll = TRUE)
}
\arguments{
\item{num.chains}{number of chains to build}

\item{n}{length of chains}

\item{ll}{log likelihood function, taking as arguments parameters p0}

\item{p0.list}{list, initial value of parameters for each chain}

\item{sigma}{covariance matrix for multivariate gaussian update kernel}

\item{lp}{function of parameters p0 giving the prior log-probability for each parameter}

\item{psuedo.ll}{indicates if ll is psuedo-likelihood, in which case it must be recalculated each time}
}
\value{
mcmc.chains object, list of chain objects
}
\description{
Build a set of MCMC chains
}
