% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{print.mcmc.chain}
\alias{print.mcmc.chain}
\title{Print convergence stats for MCMC chain}
\usage{
\method{print}{mcmc.chain}(chain)
}
\arguments{
\item{chain}{mcmc.chain}
}
\description{
Print convergence stats for MCMC chain
}
