% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{print.mcmc.chains}
\alias{print.mcmc.chains}
\title{Print convergence stats set of for MCMC chains}
\usage{
\method{print}{mcmc.chains}(chains)
}
\arguments{
\item{chains}{mcmc.chains}
}
\description{
Includes convergence stats R-hat and effective sample size (block and tail),
as computed by rstan functions \code{\link[rstan:Rhat]{rstan::Rhat()}} \code{\link[rstan:Rhat]{rstan::ess_bulk()}} \code{\link[rstan:Rhat]{rstan::ess_tail()}}.
}
