% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{prob_obs_hh}
\alias{prob_obs_hh}
\title{Probability of observed household outcome
Probability for primary infection is either computed by iteration or by simulation, depending on T.}
\usage{
prob_obs_hh(obs, prim, q, Sn = 1, Sp = 1, Thresh = -1, M = 0, it = FALSE)
}
\arguments{
\item{obs}{observed household cases. 1=infected, 0=not infection, NA=no observation}

\item{prim}{probabilities of primary infection}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}

\item{Thresh}{household size threshold for use of simulation for primary infection}

\item{M}{number of iterations for primary infection simulation}

\item{it}{whether to use iteration for secondary cases (for performance comparison only)}
}
\value{
probability of observation
}
\description{
Probability of observed household outcome
Probability for primary infection is either computed by iteration or by simulation, depending on T.
}
