% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{prob_obs_pri}
\alias{prob_obs_pri}
\title{Probability of observed household outcome given primary cases}
\usage{
prob_obs_pri(obs, pri, q, Sn = 1, Sp = 1)
}
\arguments{
\item{obs}{observed household cases (vector) NA=missed}

\item{pri}{primary household infections (vector). 0=uninfected, 1=infected}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}
}
\value{
log-probability of observation
}
\description{
Probability of observed household outcome given primary cases
}
