% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{prob_sec_obs_pri}
\alias{prob_sec_obs_pri}
\title{Probability of observation among potential secondary infections}
\usage{
prob_sec_obs_pri(obs, mis, pri, m, q, Sn = 1, Sp = 1)
}
\arguments{
\item{obs}{number of observed positive infections among potential secondary infections.}

\item{mis}{number of missed observation among potential secondary infections.}

\item{pri}{number of primary infections, which are potential sources of secondary infection.}

\item{m}{number of potential secondary household infections.}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}
}
\value{
log-probability of observation
}
\description{
Probability of observation among potential secondary infections
}
