% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{rp.jeff}
\alias{rp.jeff}
\title{Generator for samples from logistic transform of Jeffreys prior}
\usage{
rp.jeff(n)
}
\arguments{
\item{n}{number of samples to generate}
}
\description{
Generator for samples from logistic transform of Jeffreys prior
}
