% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{rp.unif}
\alias{rp.unif}
\title{Generator for samples from logistic transform of uniform prior}
\usage{
rp.unif(n)
}
\arguments{
\item{n}{number of samples to generate}
}
\description{
Generator for samples from logistic transform of uniform prior
}
