% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_genfun.R
\name{sec_prob}
\alias{sec_prob}
\title{Get secondary household infection probabilities}
\usage{
sec_prob(sec_prob_data, k, m)
}
\arguments{
\item{sec_prob_data}{pre-computed probability data (result of \code{\link[=sec_prob_init]{sec_prob_init()}})}

\item{k}{Primary infections}

\item{m}{Susceptibles (household size - k)}
}
\value{
generating function coefficients, or probabilities of infection for j=0...m
}
\description{
This uses pre-computed probabilities generated using \code{\link[=sec_prob_init]{sec_prob_init()}}
}
