% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{check.ll}
\alias{check.ll}
\title{Measure the variance of a log-likelihood estimate function}
\usage{
check.ll(p0, ll, runs = 200)
}
\arguments{
\item{p0}{parameter vector}

\item{ll}{log-likelihood estimate function}

\item{runs}{number of iterations over which to compute the variance}
}
\value{
A measure of the variance of the log-likelihood estimate function
}
\description{
Measure the variance of a log-likelihood estimate function
}
