% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{init.sig}
\alias{init.sig}
\title{Initialize covariance matrix of proposal function}
\usage{
init.sig(p.var)
}
\arguments{
\item{p.var}{vector of variances of parameter priors}
}
\value{
a nxn diagonal matrix
}
\description{
Covariance matrix is diagonal with standard deviation 1/10 sd of priors
}
