% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{lastparam.mcmc.chains}
\alias{lastparam.mcmc.chains}
\title{Get parameter values at the end of a chain}
\usage{
\method{lastparam}{mcmc.chains}(chains, n = 1)
}
\arguments{
\item{chains}{mcmc.chains}

\item{n}{number of end samples to average over}
}
\value{
list of parameter vectors of average parameters over n last samples
}
\description{
Get parameter values at the end of a chain
}
