% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.chains}
\alias{mcmc.chains}
\title{Build a set of MCMC chains}
\usage{
mcmc.chains(num.chains, n, ll, p0.list, sigma, lp = lp.jeff, psuedo.ll = TRUE)
}
\arguments{
\item{num.chains}{number of chains to build}

\item{n}{length of chain}

\item{ll}{log likelihood function, taking as arguements parameters (length of p0) and iterations it}

\item{p0.list}{initial value of parameters, for each chain}

\item{sigma}{covariance matrix for multivariate gaussian update kernel}

\item{lp}{function of two parameter vectors, should give log(prior(p1) / prior(p0))}

\item{psuedo.ll}{indicates if ll is psuedo-likelihood, in which case it must be recalculated each time}
}
\value{
list of mcmc.chain elements, containing chain, acceptance vector, estimates log-likelihood vector
}
\description{
Build a set of MCMC chains
}
