% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.marginal.lik}
\alias{mcmc.marginal.lik}
\title{MCMC Function method suitable for psuedo-likelihood}
\usage{
mcmc.marginal.lik(posterior, ll, p0, sigma, lp = lp.jeff, psuedo.ll = TRUE)
}
\arguments{
\item{posterior}{A matrix of posterior samples. Columns are parameters, rows are samples.}

\item{ll}{log likelihood function, taking as arguements parameters (length of p0) and iterations it}

\item{p0}{any paramter value, typically maximum likelihood estimate or any high likelihood estimate}

\item{sigma}{covariance matrix for multivariate gaussian update kernel}

\item{lp}{log probability density function of parameter priors}

\item{psuedo.ll}{indicates if ll is psuedo-likelihood, in which case it must be recalculated each time}
}
\value{
log marginal likelihood
}
\description{
MCMC Function method suitable for psuedo-likelihood
}
