% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{prob_obs_hh}
\alias{prob_obs_hh}
\title{Probability of observed household outcome}
\usage{
prob_obs_hh(obs, prim, q, Sn = 1, Sp = 1, Thresh = -1, M = 0, it = FALSE)
}
\arguments{
\item{obs}{observed household cases. 1=infected, 0=not infection, NA=no observation}

\item{prim}{probabilities of primary infection}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}

\item{Thresh}{household size threshold for use of simulation for primary infection}

\item{M}{number of iterations for primary infection simulation}
}
\value{
log-probability of observation
}
\description{
Probability computer by iterating or simulating primary infections:
\itemize{
\item T < 0: always iterate
\item T = 0: always simulate (M iterations)
\item T > 0: iterate for households up to size T, then simulate
}
}
