% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_obs.R
\name{prob_obs_hhs}
\alias{prob_obs_hhs}
\title{Probability of observed outcomes across multiple households}
\usage{
prob_obs_hhs(obs, prim, hh_sizes, q, Sn = 1, Sp = 1, Thresh = -1, M = 0)
}
\arguments{
\item{obs}{vector of observed cases. 1=positive for infection, 0=negative for infection, NA=unobserved individual}

\item{prim}{probabilities of primary infection}

\item{hh_sizes}{household sizes}

\item{q}{probability of secondary infection}

\item{Sn}{test sensitivity}

\item{Sp}{test specificity}

\item{Thresh}{household size threshold for use of simulation for primary infection}

\item{M}{number of iterations for primary infection simulation}
}
\value{
log-probability of observation
}
\description{
Probability computer by iterating or simulating primary infections:
\itemize{
\item T < 0: always iterate
\item T = 0: always simulate (M iterations)
\item T > 0: iterate for households up to size T, then simulate
}
}
