% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_genfun.R
\name{sec_prob}
\alias{sec_prob}
\title{Get secondary household infection probabilities}
\usage{
sec_prob(sec_prob_data, k, m)
}
\arguments{
\item{sec_prob_data}{Maximum household size for which to compute probabilities}

\item{k}{Number of primary infections}

\item{m}{Number of susceptibles (household size - k)}
}
\value{
vector of probabilities of infection for household, from 0 ... m
}
\description{
This uses pre-computed probabilities generated using sec_prob_innit
}
