% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_genfun.R
\name{sec_prob_init}
\alias{sec_prob_init}
\title{Precompute secondary infection probabilities}
\usage{
sec_prob_init(max_hh, q)
}
\arguments{
\item{max_hh}{Maximum household size for which to compute probabilities}

\item{q}{Secondary attack rate (probability of infection between any infectious and susceptible)}
}
\value{
list object containing pre-computed probabilities and max household size
}
\description{
Probabilities are generated for all household sizes from 1 to max_hh,
and for all possible numbers of primary infection from 0 to household size.
Use \code{\link[=sec_prob]{sec_prob()}} to get sepcific secondary infection probabilities.
}
